ws/pyproject.toml
[tool.poetry]
name = "alpaca-fastapi"
version = "0.1.0"
description = ""
authors = ["Doskhan Stybayev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
gunicorn = "^21.2.0"
uvicorn = "^0.27.1"
alembic = "^1.13.1"
sqlalchemy = "^2.0.28"
black = "^24.2.0"
python-dotenv = "^1.0.1"
fastapi-sqlalchemy = "^0.2.1"
promise = "^2.3"
psycopg2-binary = "^2.9.9"
pytest = "^8.0.2"
requests = "^2.31.0"
orjson = "^3.9.15"
boto3 = "^1.34.61"
pytz = "^2024.1"
logtail-python = "^0.2.10"
flake8 = "^7.0.0"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
redis = "^5.0.8"
pydantic-settings = "^2.5.2"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
asyncpg = "^0.29.0"
websockets = "^13.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
